+++ src/api/models/follow.py
@@ -1,0 +1,6 @@
+from django.db import models
+from .user import User
+
+class Follow(models.Model):
+    follower = models.ForeignKey(User, related_name='follower', on_delete=models.CASCADE)
+    followed = models.ForeignKey(User, related_name='followed', on_delete=models.CASCADE)
